Move setting servername to MainEventProcessor.
