+++ sentry/api/sentry.api
@@ -695,6 +695,7 @@ public class io/sentry/SentryOptions {
 	public fun getTracesSampler ()Lio/sentry/SentryOptions$TracesSamplerCallback;
 	public fun getTransport ()Lio/sentry/transport/ITransport;
 	public fun getTransportGate ()Lio/sentry/transport/ITransportGate;
+	public fun isAttachServerName ()Z
 	public fun isAttachStacktrace ()Z
 	public fun isAttachThreads ()Z
 	public fun isDebug ()Z
@@ -704,6 +705,7 @@ public class io/sentry/SentryOptions {
 	public fun isEnableSessionTracking ()Z
 	public fun isEnableUncaughtExceptionHandler ()Z
 	public fun isSendDefaultPii ()Z
+	public fun setAttachServerName (Z)V
 	public fun setAttachStacktrace (Z)V
 	public fun setAttachThreads (Z)V
 	public fun setBeforeBreadcrumb (Lio/sentry/SentryOptions$BeforeBreadcrumbCallback;)V
@@ -804,11 +806,6 @@ public final class io/sentry/SentryTransaction : io/sentry/SentryBaseEvent, io/s
 	public fun startChild (Lio/sentry/SpanId;Ljava/lang/String;Ljava/lang/String;)Lio/sentry/Span;
 	public fun startChild (Ljava/lang/String;Ljava/lang/String;)Lio/sentry/Span;
 	public fun toSentryTrace ()Lio/sentry/SentryTraceHeader;
-}
-
-public final class io/sentry/ServerNameResolvingEventProcessor : io/sentry/EventProcessor {
-	public fun <init> ()V
-	public fun process (Lio/sentry/SentryEvent;Ljava/lang/Object;)Lio/sentry/SentryEvent;
 }
 
 public final class io/sentry/Session {
