+++ sentry/src/test/java/io/sentry/MainEventProcessorTest.kt
@@ -1,11 +1,16 @@
 package io.sentry
 
 import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.reset
+import com.nhaarman.mockitokotlin2.times
+import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.hints.ApplyScopeData
 import io.sentry.hints.Cached
 import io.sentry.protocol.SdkVersion
 import io.sentry.protocol.User
 import java.lang.RuntimeException
+import java.net.InetAddress
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -13,6 +18,7 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertSame
 import kotlin.test.assertTrue
+import org.awaitility.kotlin.await
 
 class MainEventProcessorTest {
     class Fixture {
@@ -20,21 +26,30 @@ class MainEventProcessorTest {
             dsn = dsnString
             release = "release"
             dist = "dist"
-            serverName = "server"
             sdkVersion = SdkVersion().apply {
                 name = "test"
                 version = "1.2.3"
             }
         }
-        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap(), sendDefaultPii: Boolean? = null): MainEventProcessor {
+        val getLocalhost = mock<InetAddress>()
+
+        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap(), sendDefaultPii: Boolean? = null, serverName: String? = "server", host: String? = null, resolveHostDelay: Long? = null, hostnameCacheDuration: Long = 10): MainEventProcessor {
             sentryOptions.isAttachThreads = attachThreads
             sentryOptions.isAttachStacktrace = attachStackTrace
             sentryOptions.environment = environment
+            sentryOptions.serverName = serverName
             if (sendDefaultPii != null) {
                 sentryOptions.isSendDefaultPii = sendDefaultPii
             }
             tags.forEach { sentryOptions.setTag(it.key, it.value) }
-            return MainEventProcessor(sentryOptions)
+            whenever(getLocalhost.canonicalHostName).thenAnswer {
+                if (resolveHostDelay != null) {
+                    Thread.sleep(resolveHostDelay)
+                }
+                host
+            }
+            val hostnameCache = HostnameCache(hostnameCacheDuration) { getLocalhost }
+            return MainEventProcessor(sentryOptions, hostnameCache)
         }
     }
 
@@ -259,6 +274,68 @@ class MainEventProcessorTest {
         sut.process(event, null)
         assertEquals("value1", event.tags["tag1"])
         assertEquals("event-tag-value", event.tags["tag2"])
+    }
+
+    @Test
+    fun `sets servername retrieved from the local address`() {
+        val processor = fixture.getSut(serverName = null, host = "aHost")
+        val event = SentryEvent()
+        processor.process(event, null)
+        assertEquals("aHost", event.serverName)
+    }
+
+    @Test
+    fun `sets servername to null if retrieving takes longer time`() {
+        val processor = fixture.getSut(serverName = null, host = "aHost", resolveHostDelay = 2000)
+        val event = SentryEvent()
+        processor.process(event, null)
+        assertNull(event.serverName)
+    }
+
+    @Test
+    fun `uses cache to retrieve servername for subsequent events`() {
+        val processor = fixture.getSut(serverName = null, host = "aHost", hostnameCacheDuration = 1000)
+        val firstEvent = SentryEvent()
+        processor.process(firstEvent, null)
+        assertEquals("aHost", firstEvent.serverName)
+        val secondEvent = SentryEvent()
+        processor.process(secondEvent, null)
+        assertEquals("aHost", secondEvent.serverName)
+        verify(fixture.getLocalhost, times(1)).canonicalHostName
+    }
+
+    @Test
+    fun `when cache expires, retrieves new host name from the local address`() {
+        val processor = fixture.getSut(serverName = null, host = "aHost")
+        val firstEvent = SentryEvent()
+        processor.process(firstEvent, null)
+        assertEquals("aHost", firstEvent.serverName)
+
+        reset(fixture.getLocalhost)
+        whenever(fixture.getLocalhost.canonicalHostName).thenReturn("newHost")
+
+        await.untilAsserted {
+            val secondEvent = SentryEvent()
+            processor.process(secondEvent, null)
+            assertEquals("newHost", secondEvent.serverName)
+        }
+    }
+
+    @Test
+    fun `does not set serverName on events that already have server names`() {
+        val processor = fixture.getSut(serverName = null, host = "aHost")
+        val event = SentryEvent()
+        event.serverName = "eventHost"
+        processor.process(event, null)
+        assertEquals("eventHost", event.serverName)
+    }
+
+    @Test
+    fun `does not set serverName on events if serverName is set on SentryOptions`() {
+        val processor = fixture.getSut(serverName = "optionsHost", host = "aHost")
+        val event = SentryEvent()
+        processor.process(event, null)
+        assertEquals("optionsHost", event.serverName)
     }
 
     private fun generateCrashedEvent(crashedThread: Thread = Thread.currentThread()) = SentryEvent().apply {
